test(mir): carve the replay-plan pins out of drop_plan.rs
Adding the internal-channel routing pushed `drop_plan.rs` past its 10000-line
ceiling. The two replay-plan test modules move to sibling files unchanged, so
the ratchet holds without widening a cap.
