+++ LESSONS.md
@@ -684,6 +684,8 @@ Nuance row (table under the core):
 - **area**: `project_findings` (`hew-mir/src/lower/drop_plan/diagnostic_projection.rs`), `MirDiagnosticKind::LoweringInvariant`, `mir_diagnostic_prefix` (`hew-cli/src/diagnostic.rs`)
 - **invariant**: `DischargeAuthorityDrift` / `DischargeAuthorityMissing` / `ObligationBalanceUnverified` describe an inconsistent MIR the compiler built, never the user's program. They project to `LoweringInvariant` (code `E_MIR_ICE`); the first one per function is kept and the rest dropped — every later finding is a consequence of the same inconsistent stream (goobers `read_model::same_projection` printed 30 identical "minted while another generation is live" lines, one per temp). `ObligationUnderReleased` is reported once per `(function, mint SiteId, name)` with every unbalanced exit block listed and the mint provenances joined. Every function's findings must therefore flow through ONE `project_findings` call (`lower_function` collects them in `findings` and projects at the end); a validator that pushes straight into `diagnostics` reintroduces the fan-out.
 - **also**: `OwnerId` has a `Display` (`b<binding>#<generation>`); every verifier reason string uses it (`owner_list` / `owner_place_list` for collections). A `{owner:?}` in a `reason:`/`note:` string is a user-visible `OwnerId { binding: BindingId(4294967210), .. }` — `verifier_findings_project_to_one_internal_error_without_debug_identities` (drop_plan) and `lowering_invariant_renders_as_internal_error_without_debug_identities` (hew-cli) pin both halves.
+- **also**: the channel a finding belongs on is decided by whether the user can name its subject, not by the finding's variant. An obligation imbalance over an owner with no definition-site `Bind` (a lowering-synthesized temp), and a frozen plan entry that matches no required owner at all, both name the value out of the `OwnerId` and anchor at `SiteId(0)` — and `SiteId(0)` is a real site, so `mir_primary_site` draws a caret on the module's first construct. Both are compiler defects, so `validate_ownership_events` routes them through `unnameable_obligation_subject` to `E_MIR_ICE` instead (`an_omitted_cleanup_over_an_unnamed_owner_is_an_internal_error`, `a_plan_entry_matching_no_live_owner_is_an_internal_error`). The user register then only ever carries a source value name and its mint site, which is also what makes `(direction, function, site, name)` a sound coalescing key.
+- **also**: `project_findings` folds and suppresses, so the unprojected finding list is unrecoverable downstream. `HEW_DEBUG_CHECKED_FUNCTION=<fn>` prints `HEW_DEBUG_FINDINGS` with the whole vec immediately before projection (`lower_function`); the earlier `HEW_DEBUG_CHECKED` dump alone does NOT contain it, because `CheckedMirFunction::checks` carries only the dataflow checks passed into `seal_checked`.
 - **evidence**: goobers `src/**/*.hew` under `hew check`: 337 output lines containing `BindingId(` before F4, 0 after.
 
 ### `tmpfs-full-breaks-the-shell-tool-silently` — a full `/tmp` makes every Bash call exit 1 with no output
